Ampache: also accept authentication in "ssid" key
Until now, we only accepted the authentication token in the "auth"
key. This is the same as in Ampache.

However, for the URL of a song, Ampache passes the authentication
token in the "ssid" key. Some clients expect this key to have
this name, for example the Ampache Plugin for Rhythmbox, as shipped
with Ubuntu.

After this change, Rhythmbox works fine.

There is also a client-side fix for this problem:
<https://github.com/lotan/rhythmbox-ampache/commit/0d233e3282caea3f6a80c943bb414597a1afb733>