Ampache API: Fix playlist count in handshake response and owner in playlist response
The handshake response message has a field for the number of playlsts but the
value was always transmitted as 0. This didn't seem to matter for PowerAmpache
client, though. No other Ampahce clients with remote playlist support have been
tested.

The playlist(s) response message has a field for the playlist owner. This field was
supposed to contain the user ID but this didn't work because an undefined reference
was passed in.