Run Travis tests for different PHP versions on OC instead of NC
Changed again the Travis test config. On the previous change, I forgot that
the Nextcloud master branch no longer supports PHP5.x. Hence, it's pointless
to test this combination. On the other hand, it makes sense to test all the
different PHP versions so that there are no other variables involved.  To
achieve this, all the PHP versions are now tested on ownCloud master
instead of Nextcloud master.