Fix unhandled exception on NC13
On Nextcloud 13, the system fires the file modified hooks when it
automatically generates css files. These files have no owner, and the
handler function in the Music app was not prepared for that. Now such
files are excluded by tho hook function.

This problem was described in the comments of #629 although this was
not the actual bug reported on that issue.