Disable our embedded music player for individual shared file on NC13
Nextcloud 13 includes a built-in music player for individually shared
audio files. Furhtermore, it attaches the download link to the cover
art image where we usually link our embedded player. Disable our
embedded player if the view already contains an audio player

The disabling happens on the front-end. It would be nicer to be able to
do this already on the backend, but we cannot make distiction between
individual shared files and shared folders based only on the URL.

Fixes #630