Remove all deleted user's data
Attach to the "user deleted" hook and remove automatically all the
deleted user's data from the music database.

Note that also the behavior of resetting the DB from the command line
or by changing the music path settings was changed slightly: all the
playlists of the user are now deleted instead of just wiping clean their
contained tracks. Keeping the empty playlists around was not very
useful and it's better to have only one kind of DB erase function than
having two.