Prevent warning about incorrect count() arguments on PHP 7.2
On PHP 7.2, a warning `count(): Parameter must be an array or an object
that implements Countable at
/usr/share/webapps/nextcloud/apps2/music/db/album.php#109` got printed
to the log each time an album with no year set was fetched to build the
collection.json. On older PHP versions, such warning was not printed.
On all PHP versions, however, the call `count(null)` returns value 0 as
expected by the app logic so this was more an annoyance than a real bug.

Fixes #642