Add configuration file for PHP Coding Standards Fixer
This configuration file requires the owncloud/coding-standard package
but this was not added to dependencies in composer.json for now because
of conflict with the other currrent dev-dependencies. To use
php-cs-fixer, the composer.json has to be altered downstream or the
requried modules need to be installed by other means.

Refs #643