Fix a compatibility issue in files-music-player
Specifying default argument is an ES6 feature, probably causing
problems for old IE versions. Also, all we wanted to do was to make the
argument optional, and this is possible without default arguments: the
argument will just be 'undefined' if it is not given.

Also, jshint nagged about using == instead of ===.