FIx collection being erased even if the collection path is set to its original value
In the settings view, if user starts chaning the music collection path
and then selects the same path as was used previously, nothing is
supposed to happen. However, the logic was flawed as it compared the
previous path containing a trailing '/' to the selected path with no
traling '/'. Hence, the frontend always evaluated the path as changed,
and the user was forced to rescan the collection.