Allow resetting the music library from the settings view
There is now a button "Reset music collection" in the settings view.
This has the same effect as the `occ` command `music:reset-database
USERNAME`.

The file settingsviewcontroller.js was altered to use gettext mechanism
for translated strings instead of `t('music', 'Some string to
translate')`. Also the latter did translate the sting if translation
was available, but it did not get extracted to the .pot file when
running `make l10n-extract`.

Resolves #302
Refs #627