Fix oddities when music collection reset or path changed during scanning
Resetting the collection or changing the collection path during
scanning could lead to situation where the collection contained some
tracks which shouldn't be there. This was because one step of scanning
takes some seconds and, by default, the server serves several requests
in parallel.

Fixed by making the related API services keep the session open for the
whole duration of the service. This prevents executing other services
of the same session simultaneously and acts as kind of mutex: if such
service is running when a new request arrives, the request will wait in
a queue until the blocking service is ready.