Increase the size of the INCREMENTAL_LOAD_STEP in the albums view
As we nowadays have the custom track list directive, placing albums on
the screen is much lighter operation and can be made in bigger batches.

It was also tested if the whole "incremental load" scheme could be
ditched now. The UI would work without it even on very large
collections but the incremental loading still adds value by keeping the
UI responsive during view loading. Also, it doesn't seem to have
significant effect on the total time taken to load and show the albums
view.