Use the custom tracklist directive also in the "All tracks" view
The "All tracks" view has been separated from the "Playlist" view, now having
its own controller and php template. The "All tracks" view now utilizes the
same custom tracklist directive as the "Albums" view, making it load much
faster on large collections.

The tracklist directive has been modified a bit to enable using it in
different contexts. Mainly, some logic has been moved out from the directive
and is now provided by the host controller. These dependencies are injected
in the directive initialization.

It would probably be possible to utilize the tracklist directive also with
the custom playlists. That, however, would take some more effort because of
the more dynamic nature of those lists: it is possible to reorder and remove
tracks on the fly on those lists. Anyway, that is outside the scope of this
modification.