Make the tracklist items on "All tracks" adjust their width to content
Clicking in the empty space after the track name no longer starts playing
that track. Also, no pointer cursor or play icon is shown when hovering over
the empty area after the track name.

The same solution of using "display: table" cannot be used in the "Albums"
view because it would mess up truncating the tracks names to columns. On the
other hand, the problem is much less severe in the Albums view where the
track lists are narrow and the amount of empty space after track name is
limited.