Do not use requestAnimationFrame in the tracklist directive
Calling render function when the browser draws the next frame didn't
seem to bring any added value compared to calling it immediately upon
change on watched properties. However, it broke the compatibility with
IE9 which does not have the requestAnimationFrame function. Note that
the layout is still pretty broken on IE9 but at least the tracks can be
shown again now.