Fix the "Show all XX tracks"/"Show less" links on all versions of IE
Internet Explorer does not support assigning class values directly to
the classList property; it should be used only like
element.classList.add('some-class');
To assign multiple space-separated classes, the property className can
be used.

The album track lists were broken on all versions of IE. It was tested
that this commit fixes the issue on IE9, IE10, and IE11. The app layout
on IE9 is still quite broken in other ways, but IE10 and IE11 now work
perfectly.