Fix a small layout issue
The "Show all XX songs ..." list item had a bit smaller height than the
normal track items on the track list of an album. Hence, the height
taken by the whole track list was a bit smaller for collapsed album
with 7+ tracks than it was for an album with 6 tracks. This could cause
the albums to be laid out in an unexpected order: when artists had more
than one row of albums, some rows could start from column other than
the left-most one (i.e. visually there were empty album slots in the
middle of the album list of an artist).

Technically, this happened because the "Show all XX songs ..." list item
has the "muted" class and one of the rules applied on this class is
"display: inline;". This had to be overridden with another rule.