Follow-up fix for the Firefox scrollbar issue #631
After the previous fix  36e77d3, the unnecessary horizontal scrollbar could
still be shown on Firefox with certain browser zoom levels.

The scrollbar appeared because the width of the controls pane still ended up
being wider than the #app-view element. Now the controls pane width adjusting
logic was changed so that it directly follows the width of #app-view instead
of trying to calculate its width from the window width and some modifiers.