Store collection.json data to the file-backed ICache instead of the DB-backed Cache
On large music libraries, the size of the cached collection.json data may be
several megabytes. Depending on the DB configuration, this may be too large
an object to be inserted to the database. To prevent such problems, the json
data is now cached to a file using the ICache interface provided by the core,
instead of using our own DB-backed Cache class.

The downside of ICache is that the interface enables accessing only the cache
of the logged-in user. However, we need to be able to invalidate the cache
also in cases where the affected user is not logged in. Hence, we store the
hash of the collection to our own DB-backed Cache and consider the data in
the file-backed cache to be invalid/out-of-date if the hash is not available.

refs #588