Enable client-side caching of collection.json
Loading the collection from the server now happens in two parts:

First, the client calls `POST prepare_collection` which causes the server to
build the collection.json if it is not already cached. This endpoint returns
the hash of the collection as result.

Second, the client calls `GET collection?hash={hash value}`. The result of
this request is the actual collection.json data. In case the provided hash
value matches the hash of the returned data, the server adds the cache
headers to the HTTP response, instructing the client to cache the result for
3 months.

Refs #588