Use relative instead of absolute URLs to play tracks
Previously, the front-end requested webdav URL from the back-end and
the latter returned an absolute URL which was then used for playing.
This caused problems with some untypical proxy configurations, as it is
possible that the protocol (HTTP/HTTPS) seen by the back-end is not the
same as the one used by the front-end. Attempting to play content over
HTTP on a page otherwise using HTTPS is blocked by browsers for security
reasons.

Now the back-end returns only a relative path of the file. It's up to
the front-end to build the full webdav URL out of this. Furthermore,
the URL built by the front-end is also relative since there is no
reason for it to be absolute. The solution is inspired by the one used
in the Files app.

Fixes #650