Fix bottom of the album title being cut off on mobile layout
In case the album title was long enough to be split to two lines on
mobile layout, the lower part of the characters on the second line were
cut off. This was especially visible on Firefox but letters like 'j'
and 'g' were cut also on Chrome.

This was because the line spacing was so sparse that the two lines
didn't really fit the 50 pixel height reserved for the header. The
spacing was tightened and the vertical alignment was changed from
"middle" to the default value "baseline" as this aligned both 1-line and
2-line titles better with the cover art on the left.