Fix controls bar on the combination of NC13 and old browser
In Nextcloud 13, the core defines the positions of controls bar as
"sticky". This mode is not supported on any version of IE or on older
versions of Chromium, at least. Also, in case of the controls bar, this
position mode didn't seem to bring any added value compared the
traditional "fixed" position.