Use tablet or mobile layout in the app-view if expanded side-bar makes it too narrow to show the normal layout
The @media queries used in mobile.css and tables.css can be used to
select css rules only according to the screen size. As we now need to
select the same rules depending on the #app-view width, most of the
rules have been changed to use classes .mobile and .tablet instead of
the @media queries. These classes are set for the parent element
#app-content dynamically using javascript.