Anchor the alphabet navigation to #app-view instead of the window
Previously, the alphabet navigation had a fixed position relative to
the browser window. This position was selected so that it didn't
overlap with the full-size scrollbar used e.g. in Firefox. In addition,
the position was pushed to left by the width of an expanded sidebar.
This logic was bad, because in case of a narrow window, the sidebar
intentionally overlaps/shadows the scrollbar and part of the #app-view.
It looked odd and caused ugly overlap when the alphabet navigation was
treated differently than the scrollbar.

Now the logic has been changed so that the alphabet navigation is
aligned to the right edge of the #app-view element. The #app-view is
automatically shrunk to exclude the space taken by the scrollbar, so
there's no overlap between scrollbar and alphabet navigation. As an
added benefit, the placement of the alphabet navigation now looks
better on browsers with narrow scrollbar (Chrome, many mobile browsers).