Ensure the 'details' button shows also with long track titles
In the grid layout used on wide screens, the title is truncated so that
there's space left after the title to show the button. The "display:
table" css rule was used for this.

In the All Tracks view as well as in the narrow screen layouts of the
Albums view, the track titles are now allowed to span multiple lines.
In these cases, we can not use the same solution as in the grid layout
because it requires settings max-width for the title in pixels and we
want to utilize the whole screen width.