Make the controls bar layout even more compact on mobile portrait mode
There were some excessive margins between the control elements,
especially if the layout was used because the details pane narrowed the
view (as the layout used to reserve space for the navigation pane
open/collapse button which is not visible in this case).

The playback controls, album art, and song info texts now use
"position: relative" instead of "absolute". This makes the responsive
layouting easier as we do not have to define the left position manually
for each possible mode.