Take the details pane into account when centering the floating notes
The notes like "New files to scan" or "No audio files found" have a
fixed position in the middle of the app-view. Now it's taken into
account that the right edge of the app-view may be moved by the opened
details pane. Without this change, the note placement seemed
off-balance.