Refactoring: Split the monolithic main.php to more coherent pieces
Each section having its own ng-controller now has also own .php file
under templates/partials: controls, navigation, and sidebar.

Also, the #sm2-container element was removed as it didn't seem to serve
any purpose. According to the comment, it was there to give the user a
chance to unblock Flash but we are using soundmanager2 in the
"html5only" mode. Hence, Flash should never be used.