Bundle also the *.css file into single app.css
The main motivation for this was that IE9 has a low maximum limit of 31
for the number of distinct style  files on a single page. This limit was
reached when new css files were added in 94a2b5fb1, causing severe
layout issues (apparently, styles of angular got discarded and at
least ng-show directives stopped working).

Reducing the amount of transfered files will also save a tiny bit of
server resources.