High-light the drop-target playlist in navipane differently than the active view
When dragging a track on a playlist in the navigation pane, the
background of the said navigation pane item now turns light gray. This
is much easier to spot than the previous changing font color.

Also, having a distinct class for the drag-hovering fixes a related bug:
reordering tracks of a playlist with drag-and-drop used to remove the
highlighting of the active playlist in the navigation pane.