Refactor #app-view resize handling
Utilize the jquery.resize plugin to react directly to resize of the #app-view
element instead of observing the window resize events and other actions which
may cause the #app-view to be resized. This simplified the logic and made the
previously applied artificial delays unnecessary.

The jquery.resize was wrapped with a custom directive resize-notifier. The
directive can be attached to an element and it broadcasts resize events
through the $rootScope whenever the said element changes size.