Fix "All tracks" view showing duplicate track lists after reload
The "All tracks" view showed multiple instances of the tracklist
element if the view was kept open during scanning and then the
collection was reloaded with the reload button or automatically. The
problem vanished after navigating to another view and back.

This bug was probably introduced by the refactoring commit 79ac91b4a
where <track-list> element was replaced with <ul> element in the DOM.
Apparently, the <track-list> element had got removed automatically by
ng-if but the replacement <ul> element does not do the same. It has to
be removed manually on $destroy event.