Do not subtract the scroll bar width from embedded player on share pages if not necessary
On the file and folder share pages, we can use parent element which
already subtracts the scroll bar width if and only if necessary.

The same is not true when the player is embedded to the normal Files
app folder view. There, the best we can do is to subtract the scroll
bar width ourselves, regardless of if the scroll bar is visible or not.
We *could* manually check if the scroll bar is currently visible but
this may change when the user navigates to another folder and there
doesn't seem to be easy way to detect this.