Make the clients cache the cover images already on first load
Previously, the Music app had to be loaded twice to make the client
cache the loaded cover images. Now they are cached already after the
first load. This is achieved by using redirection responses: When the
client requests cover through a non-cacheable URL like
/api/album/{id}/cover, the server redirects the client to a cacheable
URL like /api/cover/{hash}.

This may slightly slow down the first load of the cover images but this
is more than made up on the second load.