Fix tracklist directive updating classes only when 'playing' changes
The directive tried to watch also the variable currentTrack but this
was done incorrectly: the variable belong to MainController scope and
not to the rootScope. Hence, the directive got no event when the
'currentTrack' changed and the play/pause icon in front of the current
track moved with long delay, only once the rootScope variable 'playing'
changed state.