Enqueue the whole collection when track clicked in Albums view
Previously, clicking a track in the view enqueued the album in
question. The change influences what happens after the playback reached
the last track of the album. It also changes the meaning/scope of the
Repeat and Shuffle modes.

If the user now wants to play only a single album, she has to click the
album name or cover.