Fix UI jitter on playlist create/rename forms
The positioning of the text input field did not match exactly the list
item it replaced. Hence, it looked like the text jumped up by 2 pixels
when the text input was opened.

To some degree, this depends on the browser zoom level. But at least
with 100% zooming, the UI now looks more stable.