Refactor the collection creation
The purpose of this was to split the rather ugly function to two more
coherent pieces, and to make part of the logic potentially reusable on
the Ampache API.

The refactoring decreased the performance of the function by a
negligible amount: on a large test collection of 68k+ tracks, the time
taken to build the collection increased from 20.0-20.3s to 20.3-20.5s
(i.e. by around 1%).