Prevent flooding HTTP requests from files-music-player
Similarly as in the full Music app, there is now a small delay when
starting to play a new track. This prevents DOSsing the server when the
next/previous button is clicked rapidly and repeatedly.

To indicate the user that something happened when the next/previous
was clicked, the embedded player now immediately shows the file base
name on the "artist slot" of the player. This is replaced by the
actual artist name once the information is received from the server.

refs #641