Define highlight colors to be somewhat compatible with themed bg colors
Instead, of solid light-gray color, the highlights now have black color
with lot of transparency. Hence, they don't totally screw up UI on
themes with dark background color and light font color. However, the
highlights are quite invisible on such themes but this is a much
smaller problem.

On the default theme, the highlights look pretty much the same as
before.