Simpler and all-around better implementation for Playlist in files-music-player
The file list handling was implemented on unnecessarily low level and
the logic had several problems and shortcomings. Worst of all, there
was some XML parsing issue which prevented the feature from working on
Edge or any version of IE. Chrome and Firefox worked although Firefox
logged an error to the console.

Essentially, the Playlist module tried to solve a problem which is
already solved by the core. It was pointless to try to fetch the folder
contents via WebDAV because the contents as seen by the Files app are
already available via the context reference which is passed to the file
action handler "onPlay".

In addition to fixing the compatibility with Edge and IE, the new
implementation has the following benefits:
- The playing works correctly also in non-folder views of Files:
Recent, Favorites, Shared with you, Shared with others, Shared by link,
Tags
- The playback order matches the view order of the files (at the time
the playback was started). Previously the order was always ascending
alphabetical and ordering of names with multi-digit numbers differed
from the view order.
- The folder contents are not loaded separately from the server, saving
one HTTP request and making the next/previous buttons available
immediately and not with a delay.

refs #641