Avoid huge `IN (?,?,?,...,?)` SQL statement when fetching all albums
This seems to be a tiny bit more efficient and avoids the problem of
too many host parameters when fetching a huge library from SQLite as
explained in #239.

Performance measurements:
The (artificial) test library contained 59651 tracks and 1584 albums.
Tests were run on RaspberryPi 3 with MySQL, ngingx, and PHP 7.0.
The time taken by `api/prepare_collection` before the change was 22.0
... 22.5 seconds. After the change it was 20.3 ... 20.7 seconds.