Fix the most critical layout issues on NC14
The application layout is now fully usable also on Nextcloud 14 beta and
RC versions. There are still glitches but nothing show-stopping.

This has been achieved by including style definitions which used to be set in
the core but are no longer in NC14. Defining these in the application
shouldn't affect the layout on older cloud versions.

refs #660