Fix the scroll links to work also on Nextcloud 14
On Nextcloud 14, the whole window is the scroll container instead of
just #app-content scrolling. This required changes to the the scroll
link logic.

Also the alphabet navigation links were changed to scroll through the
javascript function scrollToItem in the maincontroller instead of using
the custom angular directives du-scroll-container and du-smooth-scroll.
This way, it's easier to have different scroll logic for NC14 and other
cloud versions.

refs #660