Allow a bit larger clock deviation on Ampache handshake
In the Ampache handshake, it was previously checked that the timestamp
supplied by the client in not more than 100 seconds ahead of the system
time of the server. In case the client is using its own system clock to
create the timestamp and the client and/or the server does not
synchronize its time from the network, it could quite easily happen
that the local times of client and server differ by more than 100
seconds. The handshake failed in such cases.

Now we allow the the timestamps to differ at maximum by 10 minutes.
That shouldn't compromise the security in any significant manner but
will make the handshake more robust against small clock deviations.

refs #60