Set playlist name input element width according to #app-navigation width
On Nextcloud 14, the #app-navigation pane has width of 300px instead of
the previous 250px. The width of the text input elements used to create
or rename a playlist is now set according the available space.
Previously, it was hard-coded to 166px which looked silly on NC14. The
layout of the "New Playlist" row also used to break if scroll bar was
shown on #app-navigation.

Also, the layout of the confirm/cancel buttons of the said text input
elements was altered slightly. The confirm button is now on the left
and the cancel button on the right on the "New Playlist" row. This way,
the confrim button location is the same on "New Playlist" and when
renaming a playlist. If the playlist name is empty, the confirm button
is now disabled instead of hiding it completely.

refs #660