Use test data frozen to github in integration tests
The test content used to be downloaded from jamendo.com. However, the
metadata of the tracks there seems to change every now and then which
has broken the tests multiple times. E.g. the artist "Pascal Boisseau"
has changed exact formatting of the artists name multiple times. Hence,
the test data is now frozen to github.

To save disk space and bandwidth, all the test tracks there have been
cut to ~1-3 seconds (full length tracks couldn't be uploaded to github
in one zip because of the file size limits). From the point of view of
the tests, the audio contents of the files is not interesting.

The expected results of the integration tests have been updated to match
the file versions uploaded to github.