Fix the #controls position when collapsible navigation pane is open
In the mobile layout used on narrow windows, the #app-navigation pane is
collapsed by default. When it is expanded, the #app-content is shifted
right with css transformation. This has the side effect of changing how
elements with "position: fixed" behave. In this case, the #controls
pane got misplaced because its position was set in javascript without
taking the afore mentioned side effect into account.

The positioning of #controls pane is now fixed by using only css for
it. The differences between NC14+ and older cloud versions are now
handled by dynamically adding an extra class to the #control element on
NC14+.