Do not check for removed image files when a folder is removed
This potentially heavy operation was mostly unnecessary because of
the way the image files are used as album covers: An image file may be
used as cover only for album whose tracks are found from the same
folder. After removing a folder, there are no such tracks left.

There's one special case, where the previous logic was not pointless:
If tracks of an album are scattered to many folders, then the album may
still exist after the folder containing its cover art is removed. These
untypical cases get eventually cleaned up by the background maintenance
task, so this should not be a big deal.

refs #664