Fix a few actual issues found by Scrutinizer
The most serious of these was the one in ApiController. However, even
that was in a branch of code which is unlikely to be executed: The
shrunk down cover image would need to be too large (>100 KB) to store
it into cache. This limit should be exceeded only in very pathological
cases.