Declare support for PHP 7.3
This is an experimental change since I don't have a development
environment with PHP 7.3. The support must be declared to be able to
run the PHP 7.3 tests on Travis.

There shouldn't be much backwards compatibility issues between PHP 7.2
and 7.3 so there's a good chance that Music app works just fine on 7.3.