Do not rely on TrackBusinessLayer in Migration/PlaylistCleanup
Especially because of lack of proper documentation, it is not clear if it is
safe to call external modules from the cleanup step code. The PostMigration
step is apparently run before installing the new application files, which
could mean that call to TrackBusinessLayer would be routed to the previous
version of the said module. And as that could be one of many different
versions, guaranteeing always sensible operation would be difficult. Hence,
it's better to modify the database directly from the cleanup module instead
of using external classes.