A cleaner solution for the dark theme support
The changes needed by the theme now all happen in css. Only a small
piece of javascript is needed to detect the dark theme and set
dynamically a class for the #app element. The css can then set different
rules when this class is present.

Small changes were made on the html to make it easier (or possible) to
select the relevant elements within css.

ref #688