Properly handle accented characters on alphabet navigation
It's now possible that scroll target for some alphabet is an item
starting with an accented version of that alphabet.

Also, there's now '…' as the last alphabet navigation link. This
scrolls to the first item after the items starting with Z. On some
languages written with Latin alphabet (e.g. Finnish, Swedish),
there are some extra alphabets after Z. Furthermore, names written e.g.
in Cyrillic or Arabic characters are sorted after Z.