Jump to the beginning of the current track with the "previous" button unless already at the beginning
The "previous" button now has a double function: pressing it once during
track playback will jump to the beginning of the current track. Pressing it
while the track is already in the beginning will jump to the previous track
(as before). "At the beginning" is here defined as within 2 seconds from the
track start.

The change has been made both for the actual Music app and for the embedded
player used in the Files app.

refs #691