Allow playing everything in "folder order"
Pressing "play" icon in front of the "Folders" link on navigation pane
now plays tracks in the order they are present in the folders view.
Also, clicking any track in the folders view enqueues all the tracks
from the view instead of a single folder. This is analogous with the
behavior of the albums view.