Folders view: Fix getting the names and paths for shared and mounted folders
Previously, the shared files and folders were shown with paths present in the
file owners file system. This made no sense for the share recipient as she
sees those files with different path in her file system. This was also a
breach of confidentiality as the original path of a shared file may contain
information which the sharing user does not want to share, and doesn't
anticipate to be shared.

Also, externally mounted folders were not shown with the full path.