Fix "jump to previous" not always working while track was loading
Jumping to the previous track did not work in the following scenario:
1. Listen to a track for more than 2 seconds
2. Jump to the next track with the "Next" button
3. Jump back with the "Previous" button while the new track is still in
"Loading..." state

This was because on step 3, the play time counter was still in the
position where it was left on step 1. Hence, "Previous" button just
tried to jump back to the beginning of the current track (i.e. the
loading track). This had no effect and the counter didn't get reset,
either.