Allow getting the collection.json even if some album has invalid album artist
Previously, creating collection.json failed on unhandled exception if some
album in the DB referred to an album artist which did not exist. Now such
albums are just omitted from the collection.

Also,  albums with invalid album artist are now removed from the DB as part
of the background cleanup task. This enables fixing the DB by scanning again
the tracks of the broken albums.

refs #710