Tweak the order of executing the background cleanup steps
Attempt to make the integration test succeed again. Also, the target is
that all the obsolete and broken entries in the DB could be removed on
one execution. We now support two different cascaded removal cases:

1. Track with invalid file removed => album becomes obsolete and is
removed => artist becomes obsolete and is removed

2. Album is removed because it is corrupt by way of referring to
non-existing artist => All the tracks from the album are now orphaned
and are removed, as well