Subsonic: Provide actual music library data through the API
Note that we abuse the Subsonic API a bit by providing the library
organized by the metadata, although we use the API methods which are
intended to provide the data as organized by the filesystem
directories. Starting from Subsonic API version 1.8, there are distinct
methods which are supposed to return the data organized by metadata.
However, we are currently targeting the API version 1.4. Quite many
client applications seem to support only first versions of the API.