Lazy-load getID3 library to avoid interfering with AudioPlayer occ scan
For some reason, our ExtractorGetID3 class gets auto-loaded when
running any occ commands. When this happened, also the getID3
library got loaded. This interfered with the occ scan command of
AudioPlayer, as it depends on incompatible version of getID3 compared
to the one which we are shipping.

This change solves the problem by postponing the loading of getID3
library until it's actually used by the Music app.

refs #715