Adapt to the getID3 API change
Previously, we were using a modified version of getID3 library but now
the official version 1.9.17 includes the alternative of giving the file
to analyze as a file pointer, and there's no need for the monkey
patching. The order of the function arguments is different in the
official version compared to our self-patched one.