Subsonic: List only album artists as first level folders
Previously, all the artists were presented on the first level of the
music library. This included those artists who have no own albums,
but just perform some tracks on albums of other artists or on a
compilation album. This could make browsing the library a bit
cumbersome if there were many such artists.