Subsonic: Use the same generated passwords as with Ampache
The Subsonic API now has a proper user access control. The same table of
generated passwords is used for this as with the Ampache API.

Note that, in this DB table, the passwords are stored as sha256 hashes.
This is safer than storing the original passwords, but one of its
implications is that we can never support the token-based authentication
introduced in the Subsonic API version 1.13.0 as "recommended method".
Currently we are targeting the API version 1.4.0, and even the most
recent API versions still allow the traditional authentication, so this
might not be a significant issue.

On the other hand, there has been discussion on the Airsonic server
github, that it could be better to drop their support for the
token-based authentication so that plain-text passwords would not have
to be stored.