Subsonic: Support browsing by folders
We now publish two "top-level folders" in the Subsonic API: "Artists"
and "Folders". The first lists the artists according the metadata and
the second shows a directory tree starting from the music collection
path configured in the user settings.

The "Artists" selection is also the default if the client doesn't
specify the folder. According to the API specification, this should
mean "All folders", but that doesn't really make sense in our case
where we show the same tracks in both top-level folders, just ordered
differently.