Fix one actual bug and many smaller issues found by Scrutinizer
- The function PlaylistBusinessLayer::getPlaylistTracks accidentally
allowed obtaining track entities not belonging to the given user (in
case those entities would somehow end up on the playlist). That was
because TrackBusinessLayer::findById was accidentally called with
invalid user id, which got interpreted as "all users".

- Scrutinizer wants to have the 'use' statements for all used classes,
even if the class name is mentioned only in the function comment with
keywords @param or @return.

- There were some inconsistencies in the declared types in some
function comments.