Subsonic: Add the API method getUser
The DSub client app seems to call this API method constantly, even
though at least the version 13.0.5.2 of the app doesn't seem to use the
return value for anything as long as the declared API version is 1.4.
If the API version is incremented to 1.8, then DSub shows the data
retrieved with getUser in its "Admin" view, but then DSub and other
clients would expect to find a few other API methods, as well.